+++ src/App.tsx
@@ -1,9 +1,9 @@
+import classnames from "classnames";
 import * as React from "react";
+import { NavLink, Redirect, Route, Switch } from "react-router-dom";
 import InfoScreen from "./InfoScreen";
 import ScreenDefense from "./ScreenDefense";
 import ScreenOffense from "./ScreenOffense";
-import TabItem from "./TabItem";
-import TabContainer from "./TabContainer";
 
 const ScreenPokedex = React.lazy(async () => {
   return await import(
@@ -12,6 +12,37 @@ const ScreenPokedex = React.lazy(async () => {
     "./ScreenPokedex"
   );
 });
+
+interface TabProps {
+  title: string;
+  url: string;
+}
+
+function Tab(props: TabProps) {
+  return (
+    <NavLink
+      to={props.url}
+      className={classnames([
+        "no-underline",
+        "pv2 ph2 f5",
+        "no-outline tab-bottom-focus",
+        "b bn",
+        "br--top br4",
+        "bg-transparent",
+        "hover-black-90",
+        "black-50 bottom-border-thick",
+      ])}
+      activeClassName={classnames([
+        "black bottom-border-thick-current",
+        "no-pointer",
+      ])}
+    >
+      {props.title}
+    </NavLink>
+  );
+}
+
+Tab.displayName = "Tab";
 
 export default function App() {
   const [defenseParams, setDefenseParams] = React.useState("");
@@ -25,36 +56,47 @@ export default function App() {
             Pokémon Type Calculator
           </a>
         </h1>
-        <TabContainer>
-          <TabItem
-            url={`/offense${offenseParams}`}
-            name="offense"
-            title="Offense"
+        <div>
+          <div
+            className={classnames([
+              "flex justify-center",
+              "bg-white",
+              "bb TabBarBorder b--black-20",
+              "pt3",
+            ])}
           >
-            <ScreenOffense setOffenseParams={setOffenseParams} />
-          </TabItem>
-          <TabItem
-            url={`/defense${defenseParams}`}
-            name="defense"
-            title="Defense"
-          >
-            <ScreenDefense setDefenseParams={setDefenseParams} />
-          </TabItem>
-          <TabItem
-            url={`/pokedex${pokedexParams}`}
-            name="pokedex"
-            title="Pokédex"
-          >
-            <React.Suspense
-              fallback={<div className="Spinner center mt4 f2" />}
-            >
-              <ScreenPokedex setPokedexParams={setPokedexParams} />
-            </React.Suspense>
-          </TabItem>
-          <TabItem url="/info" name="info" title="Info">
-            <InfoScreen />
-          </TabItem>
-        </TabContainer>
+            <Tab title="Offense" url={`/offense${offenseParams}`} />
+            <Tab title="Defense" url={`/defense${defenseParams}`} />
+            <Tab title="Pokédex" url={`/pokedex${pokedexParams}`} />
+            <Tab title="Info" url="/info" />
+          </div>
+          <Switch>
+            <Route
+              path="/offense"
+              render={() => (
+                <ScreenOffense setOffenseParams={setOffenseParams} />
+              )}
+            />
+            <Route
+              path="/defense"
+              render={() => (
+                <ScreenDefense setDefenseParams={setDefenseParams} />
+              )}
+            />
+            <Route
+              path="/pokedex"
+              render={() => (
+                <React.Suspense
+                  fallback={<div className="Spinner center mt4 f2" />}
+                >
+                  <ScreenPokedex setPokedexParams={setPokedexParams} />
+                </React.Suspense>
+              )}
+            />
+            <Route path="/info" component={InfoScreen} />
+            <Redirect to="/offense" />
+          </Switch>
+        </div>
       </div>
     </div>
   );
